+++ doc/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 3.5 (TBA)
+* Removed: The depricated modules "ICMP exfiltration" and "DNS exfiltration" have been removed.
 * Revised: Improvement regarding identifying injection marker (i.e. asterisk) in provided options.
 * Revised: Improvement regarding shellshock module.
 * Added: Support regarding parsing target(s) from piped-input (i.e. stdin).